Replaced GazeInteractionPage with the contents of the GazePointerPage and then removed GazePointerPage. Updated icon images.
