+++ Microsoft.Toolkit.Uwp.SampleApp/SamplePages/GazeInteraction/GazeInteractionPage.xaml.cs
@@ -1,41 +1,123 @@
-// ******************************************************************
-// Copyright (c) Microsoft. All rights reserved.
-// This code is licensed under the MIT License (MIT).
-// THE CODE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
-// INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
-// IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
-// DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
-// TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH
-// THE CODE OR THE USE OR OTHER DEALINGS IN THE CODE.
-// ******************************************************************
-
+﻿using Microsoft.Toolkit.Uwp.Input.Gaze;
+using Microsoft.Toolkit.Uwp.UI.Extensions;
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Runtime.InteropServices.WindowsRuntime;
+using Windows.Foundation;
+using Windows.Foundation.Collections;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
+using Windows.UI.Xaml.Controls.Primitives;
+using Windows.UI.Xaml.Data;
+using Windows.UI.Xaml.Input;
+using Windows.UI.Xaml.Media;
+using Windows.UI.Xaml.Navigation;
+using Windows.UI.Xaml.Shapes;
+
+// The Blank Page item template is documented at https://go.microsoft.com/fwlink/?LinkId=234238
 
 namespace Microsoft.Toolkit.Uwp.SampleApp.SamplePages
 {
     /// <summary>
     /// An empty page that can be used on its own or navigated to within a Frame.
     /// </summary>
-    public sealed partial class GazeInteractionPage : Page
+    public sealed partial class GazeInteractionPage : IXamlRenderListener
     {
-        private uint _button1ClickCount = 0;
-        private uint _togglebutton1ClickCount = 0;
+        private GazeElement gazeButtonControl;
+        
+        private Rectangle enterRec;
+        private Rectangle fixationRec;
+        private Rectangle dwellRec;
+        private Rectangle repeatRec;
+        private Rectangle exitRec;
+        private TextBlock dwellCountText;
+
+        private int dwellCount = 0;
 
         public GazeInteractionPage()
         {
             this.InitializeComponent();
         }
 
-        private void Button1_Click(object sender, RoutedEventArgs e)
+        public void OnXamlRendered(FrameworkElement control)
         {
-            TextBlock_Button1.Text = $"Clicks = {++_button1ClickCount}";
+            dwellCountText = control.FindChildByName("DwellCountText") as TextBlock;
+
+            enterRec = control.FindChildByName("EnterRec") as Rectangle;
+            fixationRec = control.FindChildByName("FixationRec") as Rectangle;
+            dwellRec = control.FindChildByName("DwellRec") as Rectangle;
+            repeatRec = control.FindChildByName("RepeatRec") as Rectangle;
+            exitRec = control.FindChildByName("ExitRec") as Rectangle;
+
+            var buttonControl = control.FindChildByName("TargetButton") as Button;
+
+            if (buttonControl != null)
+            {
+                gazeButtonControl = GazeApi.GetGazeElement(buttonControl);
+
+                if (gazeButtonControl == null)
+                {
+                    gazeButtonControl = new GazeElement();
+                    GazeApi.SetGazeElement(buttonControl, gazeButtonControl);
+                }
+
+                if (gazeButtonControl != null)
+                {
+                    gazeButtonControl.Invoked += OnGazeInvoked;
+                    gazeButtonControl.StateChanged += GazeButtonControl_StateChanged;
+                }
+            }
         }
 
-        private void ToggleButton1_Checked(object sender, RoutedEventArgs e)
+        private void OnGazeInvoked(object sender, GazeInvokedRoutedEventArgs e)
         {
-            TextBlock_ToggleButton1.Text = $"Checks = {++_togglebutton1ClickCount}";
+
+
         }
+
+        private void GazeButtonControl_StateChanged(object sender, GazePointerEventArgs ea)
+        {
+
+            if (ea.PointerState == GazePointerState.Enter)
+            {
+                enterRec.Visibility = Visibility.Visible;
+                dwellCountText.Visibility = Visibility.Collapsed;
+                dwellCountText.Text = "";
+                dwellCount = 0;
+                exitRec.Visibility = Visibility.Collapsed;
+            }
+            if (ea.PointerState == GazePointerState.Fixation)
+            {
+                fixationRec.Visibility = Visibility.Visible;
+            }
+            if (ea.PointerState == GazePointerState.Dwell)
+            {
+                if (dwellCount == 0)
+                {
+                    dwellRec.Visibility = Visibility.Visible;
+                    dwellCount = 1;
+                }
+                else
+                {
+                    repeatRec.Visibility = Visibility.Visible;                    
+                    dwellCountText.Text = dwellCount.ToString();
+                    dwellCountText.Visibility = Visibility.Visible;
+                    dwellCount += 1;
+                }
+            }
+            if (ea.PointerState == GazePointerState.Exit)
+            {
+                exitRec.Visibility = Visibility.Visible;
+
+                enterRec.Visibility = Visibility.Collapsed;
+                fixationRec.Visibility = Visibility.Collapsed;
+                dwellRec.Visibility = Visibility.Collapsed;
+                repeatRec.Visibility = Visibility.Collapsed;
+            }
+
+        }
+
     }
 }
